Updates the automating rollover tutorials (#3188)
Adding more context to help orient users. For example, explaining how
time is calculated from the last rollover when moving data through
phases (and not from index creation).
I've also added a new screenshot for the index settings page of the
**Create template** wizard, to explain that that is how you tie in your
template to the policy created in the previous step.

Partially addresses tasks defined in #1566.