Add more details about ES|QL chart preservation scenarios (#3247)
This PR documents the improvements to ES|QL viz chart configuration
preservation that better take user intent into consideration. To avoid
ending up with a very long note, I created a sub section that describes
the behavior and distinguishes between 9.1 and 9.2 improvements, so
users on any of these versions know what to expect or have a minimum of
information to understand.

@stratoula let me know if I correctly outlined the various reasons why
the configuration could reset.

Closes: https://github.com/elastic/docs-content/issues/3087