Introducing Remote cluster filters in the docs (#3034)
This PR attempts to introduce and document the Remote Cluster filters
that are available in ECH and ECE.

We had some small statements in certain remote clusters docs about
needing this type of filter when network security was enabled in the
remote cluster, but we didn't have the filter documented.

Changes:
- Remote cluster filter is presented as a type of Private Connection
policy in ECH and a type of filter rule set in ECE. This is located in
Security -> Network Security docs, as filters belong there.

- "Remote clusters and network security" information rehomed to the
landing page of Remote Clusters, as it was duplicate in Remote Clusters
> ECH and Remote Clusters > ECE docs --> If we prefer the previous
approach we can use snippets and remove the content from the landing
page, but let's start focusing on reviewing the content itself.

- The notes of different use cases docs (ECE --> ECH, ECH --> ECH, etc)
have been updated to reflect the reality, and to address
https://github.com/elastic/docs-content-internal/issues/59.

---------

Co-authored-by: Alex Chalkias <34575586+alxchalkias@users.noreply.github.com>
Co-authored-by: shainaraskas <58563081+shainaraskas@users.noreply.github.com>