Variable controls in Discover (#3440)
This PR:

- Adds a section about creating variable controls in Discover, with
sub-sections on how to edit them and how to bring these controls over to
a dashboard
- Transforms certain pieces of content into content snippets that are
shared between Discover and Dashboards docs as variable controls are
used in both apps: code examples, control parameters.
- Adds the section about "how to bring Discover sessions along with
their controls to a dashboard" to the "Add controls" page in dashboards
docs as well. This content is single-sourced so only maintained in one
place.
- Re-adds a "preview" flag to the Variable controls section in dashboard
docs, that one was mistakenly removed in another PR

Find page previews in one of the comments below

Note: Tested on staging environment on Oct 10

Closes: https://github.com/elastic/docs-content/issues/3096

---------

Co-authored-by: Stratou <stratoula1@gmail.com>
Co-authored-by: Benjamin Ironside Goldstein <91905639+benironside@users.noreply.github.com>