Background search in Discover and Dashboards (#3455)
This PR:

- Repurposes the Search sessions page into Background search. The
feature being essentially the same on the UX side (a few changes still),
and using the same settings behind the scenes, we chose to update the
existing page rather than create a new one.
- Calls out that search sessions, which is a deprecated feature since
8.15 and disabled by default, are being replaced with this new
background search feature, so that search sessions users know what's
happening.
- Adds references to background search to a few places to increase its
visibility.
   - Discover overview docs on searching & filtering
   - Glossary
   - Dashboards docs as a tip for slow loading scenarios
- Removes the old search sessions example in favor of more generic but
goal-oriented small sections

The update to the Search session settings page are in
https://github.com/elastic/kibana/pull/238699 as these source files are
in another repo

Closes: https://github.com/elastic/docs-content/issues/3105

---------

Co-authored-by: Arianna Laudazzi <46651782+alaudazzi@users.noreply.github.com>