[E&A] Marks ELSER on EIS as GA (#3014)
## Overview

Related to https://github.com/elastic/search-team/issues/10758 and
https://github.com/elastic/docs-content/issues/3008

This PR:
* amends the `applies_to` tags relevant to ELSER on EIS to mark the
feature as GA
* adds a Pricing section that links to the pricing page
* removes limitations that are not relevant anymore

### Important note

**Serverless:**
After this PR is merged, the docs will state that ELSER on EIS is GA on
Serverless. **Only merge once the feature is actually released as GA on
Serverless.**

**Stack:**
If the PR is merged before the 9.2 release, the docs will show the GA
status on the stack as Planned, without referencing a specific version.
After the release, the badge will update dynamically to show the feature
as GA in 9.2.

---------

Co-authored-by: Sean Handley <seanhandley@users.noreply.github.com>
Co-authored-by: Liam Thompson <leemthompo@gmail.com>
Co-authored-by: Max Jakob <max.jakob@elastic.co>
Co-authored-by: florent-leborgne <florent.leborgne@elastic.co>