ECE version 3.8.3 and 4.0.3 docs-content updates (#3752)
This PR includes the info about system deployments on ECE 4.0.3 and
3.8.3.

It also adds some missing substitutions.

There's no change in the main substitutions as they remain the same:

```
  ece-docker-images-8: 8.18.8
  ece-docker-images-9: 9.0.8
```

version.ece bump to 4.0.3 is NOT related with this PR, that's changed in
the main bump ECE PR on docs-builder repo.

cc: @claudia-correia / @elastic/admin-docs