Add troubleshooting guide for 429 errors when using the Elastic Cloud Managed OTLP Endpoint (#3669)
This PR adds a new troubleshooting topic under that explains how to
diagnose and resolve HTTP 429 Too Many Requests errors when sending data
to the mOTLP endpoint in both Elastic Cloud Serverless and Elastic Cloud
Hosted environments.

Also includes an update to the mOTLP quickstart page that replaces the
inline “Error: too many requests” section with a link to this
troubleshooting guide.

Closes [#6054](https://github.com/elastic/ingest-dev/issues/6054)