Deprecates "Add data from Splunk" without removing section (#3808)
Fixes https://github.com/elastic/docs-content-internal/issues/403 by
removing the Add Data from Splunk page, which was deprecated in 9.0 This
was a technical preview feature, and the current docs are no longer
accurate; the functionality is no longer present starting from 9.0 per
@jamiehynds. Also removes links to that page, since we shouldn't be
directing users to a deprecated feature.

(Does not remove the "other tutorials" section that I proposed removing
in a previous PR).