Documents new RBAC for value reports (#3817)
Fixes #3720 — documents a new RBAC control that controls access to the
Value Report feature in Elastic Security (only available on serverless
security and EASE projects for now — planned for Stack v9.3).

Tweaks the value reports page to make it more generic, since previously
it was only available on EASE, and it's now available in serverless.

Although in the original ticket @stephmilovic requested that the new
RBAC control be documented
[here](https://www.elastic.co/docs/solutions/security/get-started/elastic-security-requirements),
for this draft I opted to document it directly on the Value Reports
page, since the security requirements page is more of a high-level page
that isn't focused on specific features (other than linking to other
requirements pages that are related to specific features). Instead, I
suggest we make the in-product link go to the Value Reports page rather
than the security requirements page.

Also added the value reports page to another spot in the GenAI for
security section — it's still present in the original spot in the EASE
subfolder, but now that it's available in serverless too, I wanted to
make it more findable.

---------

Co-authored-by: florent-leborgne <florent.leborgne@elastic.co>