[ON week] AI-feedback updates to Security AI docs (#3935)
Fixes #3933. This PR is based on Nastasha's and my on week project.
We've been using LLMs to conduct user research on our docs. After
iterating on our prompt and trying it with severlal LLMs, I opened this
PR because I wanted to make some initial updates to the docs in the
section I've been focusing on for user research: Security AI docs.

This PR contains a variety of edits intended to fix minor mistakes
(redundancy, imprecise language, and unanswered questions.

This PR was created using AI, mostly GPT5 and Claude sonnet 4.5. Some of
the text was drafted by AI. All of it was edited by me.

---------

Co-authored-by: Nastasha Solomon <79124755+nastasha-solomon@users.noreply.github.com>