Refine description of anomaly detection techniques (#3981)
"use a bespoke amalgamation of techniques"

This phrase is very typically interpreted as "a bunch of popsicle stick
and chewing gum features" when customers read it and this is not the
case. I suggest this simple change to clear up this perception.

I suggest replacing this with "use a combination of advanced
mathematical techniques"