Enhance and restructure AutoOps pages (#3955)
In this PR, I have:
- added a child to the **Troubleshooting** page because it had too much
info. More child pages will be added in the future.
  - made various general fixes.
- ~- moved the **Views** and **Events** pages under the **AutoOps for
ECH** section and linked to them from the **AutoOps for SM clusters**
section because they don't apply to **AutoOps for Serverless**. I'm open
to feedback if anyone has a better idea for organizing this info.~


Closes https://github.com/elastic/docs-content/issues/3220