[Integrations] Alerting Rule Template common page (#4072)
<!--
Thank you for contributing to the Elastic Docs! 🎉
Use this template to help us efficiently review your contribution.
-->

## Summary

This PR adds a new Alerting Rule Templates page that should serve as a
centralized doc page with an explanation of what the alert rule
templates are and how to use them. This page is going to be referenced
from the individual integration pages.

The current location of the page will be changed and made more visible
in a further iteration.


Relates to
https://github.com/elastic/docs-content/issues/3678#issuecomment-3561707267.


## Generative AI disclosure
<!--
To help us ensure compliance with the Elastic open source and
documentation guidelines, please answer the following:
-->
1. Did you use a generative AI (GenAI) tool to assist in creating this
contribution?
- [x] Yes  
- [ ] No  

Cursor with gpt5

---------

Co-authored-by: muthu-mps <101238137+muthu-mps@users.noreply.github.com>
Co-authored-by: Fabrizio Ferri-Benedetti <fabri.ferribenedetti@elastic.co>