Fix serverless applicability in setting up data streams docs (#4083)
Fixes #3384 and updates the page so that it can be used for both Stack
and Serverless scopes.

## Summary
I've added a bit more context around the step to create an ILM policy
included in this tutorial, and why it is not required in a Serverless
context.

I've also added tabs to the sections that list both Kibana steps and
steps using the API. Each section now also links to the corresponding
Serverless API doc page.

## Generative AI disclosure
<!--
To help us ensure compliance with the Elastic open source and
documentation guidelines, please answer the following:
-->
1. Did you use a generative AI (GenAI) tool to assist in creating this
contribution?
- [ ] Yes  
- [x] No  
<!--
2. If you answered "Yes" to the previous question, please specify the
tool(s) and model(s) used (e.g., Google Gemini, OpenAI ChatGPT-4, etc.).

Tool(s) and model(s) used:
-->

---------

Co-authored-by: Marci W <333176+marciw@users.noreply.github.com>