[Elastic Agent] Update diagnostics structure and document how to disable output status reporting (#4155)
## Summary

We have introduced beats receivers and now use the OpenTelemetry
collector to perform self-monitoring by default in 9.2.1. See
https://github.com/elastic/elastic-agent/blob/main/docs/hybrid-agent-beats-receivers.md.

This has brought in a few changes the public documentation should
include:

1. There are several new files in the diagnostics structure to document.
Probably the formatting of this could be improved but I stuck with what
was already there.
2. The output for inputs that use beat receivers will now report
themselves as degraded or failed if there is a connectivity issue. This
can be turned off in the output settings. The main reason to do this is
because there is no output status UI yet and/or users may find false
health changes annoying.

## Generative AI disclosure
1. Did you use a generative AI (GenAI) tool to assist in creating this
contribution?
- [ ] Yes  
- [x] No  
<!--

---------

Co-authored-by: Visha Angelova <91186315+vishaangelova@users.noreply.github.com>