Emphasise that OIDC is only for interactive users (#3875)
OpenID Connect is a protocol for authenticating interactive users.
However, because identity platforms that support OIDC often support
other protocols that can be used for non-interactive users, it is common
for readers to end up on the OIDC page looking for ways to authenticate
non-interactive users.

The docs attempt to correct this misconception by refering to a "web
application" but that was easy to miss. This change makes it more
explicit that OIDC is not a viable option for service authentication and
recommends JWT instead.