Add serverless changelog for December 8, 2025 (#4274)
## Summary

This PR adds the Serverless changelog for Dec 8, 2025. This week there
was no Elasticsearch Serverless release promotion, so this PR does not
include ES PRs.

Resolves https://github.com/elastic/docs-content/issues/4254

## Generative AI disclosure

1. Did you use a generative AI (GenAI) tool to assist in creating this
contribution?
- [x] Yes  
- [ ] No  

2. Tool(s) and model(s) used:

I used Cursor with the gpt-5.1 model to format the PR descriptions to
match the style of the Serverless release notes. I manually checked the
descriptions against the PRs and fixed/ rephrased the descriptions where
necessary.

---------

Co-authored-by: David Kilfoyle <41695641+kilfoyle@users.noreply.github.com>
Co-authored-by: Janeen Mikell Roberts <57149392+jmikell821@users.noreply.github.com>