Moves LLM connector guides to explore-analyze section (#4224)
## Summary

Moves the LLM connector guides from the security docs section to the new
AI section within explore-analyze. This fixes
elastic/docs-content-internal/issues/487 as part of
elastic/docs-content-internal/issues/298. The purpose is to update the
IA for these docs to be solution-agnostic, since they are helpful to
users of any solution, not just security.

This PR:
- Moves the four connector-specific guides to the recently created
section, in a new sub-section.
- Creates a landing page for the new subsection.
- Moves `explore-analyze/ai-features.md` to
`explore-analyze/ai-features/ai-features.md` (this is just cleanup from
our [previous PR](https://github.com/elastic/docs-content/pull/3768)).
- Updates links to the LLM connector guides and creates corresponding
redirects.
- Moves the mapped_pages frontmatter from the page within the security
solution docs that previously was the landing page for the connector
guides (the page should still exist as there are some security-specific
guides that were not moved) to the new landing page within
explore-analyze. I would appreciate a check on my thinking here, but my
thinking is that we should generally be sending users to the new landing
page.
- Updates the security LLM connectors landing page introduction.
- Updates Elastic Managed LLM snippet.
- Fixes a few Vale-identified sentence-level issues in the Amazon
Bedrock guide.

## Generative AI disclosure
1. Did you use a generative AI (GenAI) tool to assist in creating this
contribution?
- [x ] Yes  
- [ ] No  

Tool(s) and model(s) used:
CoPilot with GPT 4.1 to help with the redirects.

---------

Co-authored-by: Mike Birnstiehl <114418652+mdbirnstiehl@users.noreply.github.com>
Co-authored-by: florent-leborgne <florent.leborgne@elastic.co>